requirement.txt
flask_wtf
Flask-SQLAlchemy
bootstrap-flask
requests
